fix:выполнен рефакторинг и добавлены стили для мобильной версии
