+++ src/components/pages/chat-room/chat-room-page.css
@@ -7,22 +7,6 @@
   justify-content: space-between;
 }
 
-.chat-header {
-  display: flex;
-  align-items: center;
-  height: 16%;
-  background-color: rgb(211, 211, 211);
-}
-
-.chat-header h2 {
-  margin-left: 21px;
-}
-
-.user-pic {
-  margin: 15px;
-  width: 5%;
-}
-
 .chat-container {
   display: flex;
   height: 76%;
@@ -43,26 +27,6 @@
   overflow-y: scroll;
 }
 
-.chat-footer {
-  height: 8%;
-  display: flex;
-  justify-content: center;
-  align-content: center;
-  background: rgb(211, 211, 211);
-}
-
-.chat-footer input, button {
-  margin: 13px;
-}
-
-.chat-footer-text-input {
-  width: 70%;
-  border-radius: 15px;
-  border: 1px solid #fff;
-  background: rgb(243, 243, 243);
-  outline: none;
-}
-
 .new-message {
   position: absolute;
   top: 84%;
@@ -73,25 +37,19 @@
   background: #bafff042;
 }
 
-.chat-footer-button {
-  background: none;
-  border: none;
-  width: 50px;
-}
-
-
 .new-message:focus {
   border: 1px solid #ffffff;
   border-radius: 15px;
   background: #bafff0ab;
 }
-.chat-footer-image-label {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-content: center;
-}
 
-.chat-footer-image-input {
-  display: none;
+@media (max-width: 768px) {
+  .sidebar {
+    display: none;
+  }
+
+  .chat-list {
+    width: 100%;
+    overflow-y: scroll;
+  }
 }
