+++ src/components/pages/chat-room/footer/index.js
@@ -1,0 +1,3 @@
+import Footer from "./footer";
+
+export default Footer;
